requirements.txt
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
Pillow==5.1.0
PyJWT==2.3.0
sqlparse==0.4.2

